Move batchVectorizer into a shared package
