+++ modules/text2vec-openai/clients/openai.go
@@ -23,6 +23,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/weaviate/weaviate/entities/modulecapabilities"
+	"github.com/weaviate/weaviate/entities/moduletools"
+
 	"github.com/weaviate/weaviate/usecases/modulecomponents"
 
 	"github.com/pkg/errors"
@@ -98,7 +101,7 @@ func buildUrl(baseURL, resourceName, deploymentID string, isAzure bool) (string,
 	return url.JoinPath(host, path)
 }
 
-type vectorizer struct {
+type client struct {
 	openAIApiKey       string
 	openAIOrganization string
 	azureApiKey        string
@@ -107,8 +110,8 @@ type vectorizer struct {
 	logger             logrus.FieldLogger
 }
 
-func New(openAIApiKey, openAIOrganization, azureApiKey string, timeout time.Duration, logger logrus.FieldLogger) *vectorizer {
-	return &vectorizer{
+func New(openAIApiKey, openAIOrganization, azureApiKey string, timeout time.Duration, logger logrus.FieldLogger) *client {
+	return &client{
 		openAIApiKey:       openAIApiKey,
 		openAIOrganization: openAIOrganization,
 		azureApiKey:        azureApiKey,
@@ -120,20 +123,22 @@ func New(openAIApiKey, openAIOrganization, azureApiKey string, timeout time.Dura
 	}
 }
 
-func (v *vectorizer) Vectorize(ctx context.Context, input []string,
-	config ent.VectorizationConfig,
-) (*ent.VectorizationResult, *ent.RateLimits, error) {
+func (v *client) Vectorize(ctx context.Context, input []string,
+	cfg moduletools.ClassConfig,
+) (*modulecapabilities.VectorizationResult, *modulecapabilities.RateLimits, error) {
+	config := v.getVectorizationConfig(cfg)
 	return v.vectorize(ctx, input, v.getModelString(config.Type, config.Model, "document", config.ModelVersion), config)
 }
 
-func (v *vectorizer) VectorizeQuery(ctx context.Context, input []string,
-	config ent.VectorizationConfig,
-) (*ent.VectorizationResult, error) {
+func (v *client) VectorizeQuery(ctx context.Context, input []string,
+	cfg moduletools.ClassConfig,
+) (*modulecapabilities.VectorizationResult, error) {
+	config := v.getVectorizationConfig(cfg)
 	res, _, err := v.vectorize(ctx, input, v.getModelString(config.Type, config.Model, "query", config.ModelVersion), config)
 	return res, err
 }
 
-func (v *vectorizer) vectorize(ctx context.Context, input []string, model string, config ent.VectorizationConfig) (*ent.VectorizationResult, *ent.RateLimits, error) {
+func (v *client) vectorize(ctx context.Context, input []string, model string, config ent.VectorizationConfig) (*modulecapabilities.VectorizationResult, *modulecapabilities.RateLimits, error) {
 	body, err := json.Marshal(v.getEmbeddingsRequest(input, model, config.IsAzure, config.Dimensions))
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "marshal body")
@@ -191,7 +196,7 @@ func (v *vectorizer) vectorize(ctx context.Context, input []string, model string
 		}
 	}
 
-	return &ent.VectorizationResult{
+	return &modulecapabilities.VectorizationResult{
 		Text:       texts,
 		Dimensions: len(resBody.Data[0].Embedding),
 		Vector:     embeddings,
@@ -199,7 +204,7 @@ func (v *vectorizer) vectorize(ctx context.Context, input []string, model string
 	}, rateLimit, nil
 }
 
-func (v *vectorizer) buildURL(ctx context.Context, config ent.VectorizationConfig) (string, error) {
+func (v *client) buildURL(ctx context.Context, config ent.VectorizationConfig) (string, error) {
 	baseURL, resourceName, deploymentID, isAzure := config.BaseURL, config.ResourceName, config.DeploymentID, config.IsAzure
 	if headerBaseURL := v.getValueFromContext(ctx, "X-Openai-Baseurl"); headerBaseURL != "" {
 		baseURL = headerBaseURL
@@ -207,7 +212,7 @@ func (v *vectorizer) buildURL(ctx context.Context, config ent.VectorizationConfi
 	return v.buildUrlFn(baseURL, resourceName, deploymentID, isAzure)
 }
 
-func (v *vectorizer) getError(statusCode int, resBodyError *openAIApiError, isAzure bool) error {
+func (v *client) getError(statusCode int, resBodyError *openAIApiError, isAzure bool) error {
 	endpoint := "OpenAI API"
 	if isAzure {
 		endpoint = "Azure OpenAI API"
@@ -218,28 +223,28 @@ func (v *vectorizer) getError(statusCode int, resBodyError *openAIApiError, isAz
 	return fmt.Errorf("connection to: %s failed with status: %d", endpoint, statusCode)
 }
 
-func (v *vectorizer) getEmbeddingsRequest(input []string, model string, isAzure bool, dimensions *int64) embeddingsRequest {
+func (v *client) getEmbeddingsRequest(input []string, model string, isAzure bool, dimensions *int64) embeddingsRequest {
 	if isAzure {
 		return embeddingsRequest{Input: input}
 	}
 	return embeddingsRequest{Input: input, Model: model, Dimensions: dimensions}
 }
 
-func (v *vectorizer) getApiKeyHeaderAndValue(apiKey string, isAzure bool) (string, string) {
+func (v *client) getApiKeyHeaderAndValue(apiKey string, isAzure bool) (string, string) {
 	if isAzure {
 		return "api-key", apiKey
 	}
 	return "Authorization", fmt.Sprintf("Bearer %s", apiKey)
 }
 
-func (v *vectorizer) getOpenAIOrganization(ctx context.Context) string {
+func (v *client) getOpenAIOrganization(ctx context.Context) string {
 	if value := v.getValueFromContext(ctx, "X-Openai-Organization"); value != "" {
 		return value
 	}
 	return v.openAIOrganization
 }
 
-func (v *vectorizer) getApiKey(ctx context.Context, isAzure bool) (string, error) {
+func (v *client) getApiKey(ctx context.Context, isAzure bool) (string, error) {
 	var apiKey, envVar string
 
 	if isAzure {
@@ -259,14 +264,14 @@ func (v *vectorizer) getApiKey(ctx context.Context, isAzure bool) (string, error
 	return v.getApiKeyFromContext(ctx, apiKey, envVar)
 }
 
-func (v *vectorizer) getApiKeyFromContext(ctx context.Context, apiKey, envVar string) (string, error) {
+func (v *client) getApiKeyFromContext(ctx context.Context, apiKey, envVar string) (string, error) {
 	if apiKeyValue := v.getValueFromContext(ctx, apiKey); apiKeyValue != "" {
 		return apiKeyValue, nil
 	}
 	return "", fmt.Errorf("no api key found neither in request header: %s nor in environment variable under %s", apiKey, envVar)
 }
 
-func (v *vectorizer) getValueFromContext(ctx context.Context, key string) string {
+func (v *client) getValueFromContext(ctx context.Context, key string) string {
 	if value := ctx.Value(key); value != nil {
 		if keyHeader, ok := value.([]string); ok && len(keyHeader) > 0 && len(keyHeader[0]) > 0 {
 			return keyHeader[0]
@@ -280,7 +285,7 @@ func (v *vectorizer) getValueFromContext(ctx context.Context, key string) string
 	return ""
 }
 
-func (v *vectorizer) getModelString(docType, model, action, version string) string {
+func (v *client) getModelString(docType, model, action, version string) string {
 	if strings.HasPrefix(model, "text-embedding-3") {
 		// indicates that we handle v3 models
 		return model
@@ -291,7 +296,7 @@ func (v *vectorizer) getModelString(docType, model, action, version string) stri
 	return v.getModel001String(docType, model, action)
 }
 
-func (v *vectorizer) getModel001String(docType, model, action string) string {
+func (v *client) getModel001String(docType, model, action string) string {
 	modelBaseString := "%s-search-%s-%s-001"
 	if action == "document" {
 		if docType == "code" {
@@ -307,7 +312,21 @@ func (v *vectorizer) getModel001String(docType, model, action string) string {
 	}
 }
 
-func (v *vectorizer) getModel002String(model string) string {
+func (v *client) getModel002String(model string) string {
 	modelBaseString := "text-embedding-%s-002"
 	return fmt.Sprintf(modelBaseString, model)
+}
+
+func (v *client) getVectorizationConfig(cfg moduletools.ClassConfig) ent.VectorizationConfig {
+	settings := ent.NewClassSettings(cfg)
+	return ent.VectorizationConfig{
+		Type:         settings.Type(),
+		Model:        settings.Model(),
+		ModelVersion: settings.ModelVersion(),
+		ResourceName: settings.ResourceName(),
+		DeploymentID: settings.DeploymentID(),
+		BaseURL:      settings.BaseURL(),
+		IsAzure:      settings.IsAzure(),
+		Dimensions:   settings.Dimensions(),
+	}
 }
