+++ modules/text2vec-openai/ent/class_settings.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package vectorizer
+package ent
 
 import (
 	"fmt"
@@ -146,7 +146,7 @@ func (cs *classSettings) Validate(class *models.Class) error {
 		return err
 	}
 
-	err = cs.validateIndexState(class, cs)
+	err = cs.validateIndexState(class, cs.VectorizeClassName(), cs.PropertyIndexed)
 	if err != nil {
 		return err
 	}
@@ -198,8 +198,8 @@ func (cs *classSettings) getPropertyAsInt(name string, defaultValue *int64) *int
 	return cs.BaseClassSettings.GetPropertyAsInt64(name, defaultValue)
 }
 
-func (cs *classSettings) validateIndexState(class *models.Class, settings ClassSettings) error {
-	if settings.VectorizeClassName() {
+func (cs *classSettings) validateIndexState(class *models.Class, vectorizeClassName bool, propIndexed func(string) bool) error {
+	if vectorizeClassName {
 		// if the user chooses to vectorize the classname, vector-building will
 		// always be possible, no need to investigate further
 
@@ -219,7 +219,7 @@ func (cs *classSettings) validateIndexState(class *models.Class, settings ClassS
 			continue
 		}
 
-		if settings.PropertyIndexed(prop.Name) {
+		if propIndexed(prop.Name) {
 			// found at least one, this is a valid schema
 			return nil
 		}
