+++ modules/text2vec-openai/vectorizer/objects_test.go
@@ -13,9 +13,10 @@ package vectorizer
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
+
+	"github.com/weaviate/weaviate/modules/text2vec-openai/ent"
 
 	"github.com/sirupsen/logrus/hooks/test"
 
@@ -186,7 +187,7 @@ func TestVectorizingObjects(t *testing.T) {
 
 			v := New(client, 40*time.Second, logger)
 
-			cfg := &fakeClassConfig{
+			cfg := &FakeClassConfig{
 				classConfig: map[string]interface{}{
 					"vectorizeClassName": test.excludedClass != "Car",
 					"type":               test.openAIType,
@@ -202,8 +203,9 @@ func TestVectorizingObjects(t *testing.T) {
 			require.Nil(t, err)
 			assert.Equal(t, []float32{0, 1, 2, 3}, vector)
 			assert.Equal(t, []string{test.expectedClientCall}, client.lastInput)
-			assert.Equal(t, test.expectedOpenAIType, client.lastConfig.Type)
-			assert.Equal(t, test.expectedOpenAIModel, client.lastConfig.Model)
+			conf := ent.NewClassSettings(client.lastConfig)
+			assert.Equal(t, test.expectedOpenAIType, conf.Type())
+			assert.Equal(t, test.expectedOpenAIModel, conf.Model())
 		})
 	}
 }
@@ -215,13 +217,13 @@ func TestClassSettings(t *testing.T) {
 	}
 	tests := []testCase{
 		{
-			cfg: fakeClassConfig{
+			cfg: FakeClassConfig{
 				classConfig: make(map[string]interface{}),
 			},
-			expectedBaseURL: DefaultBaseURL,
+			expectedBaseURL: ent.DefaultBaseURL,
 		},
 		{
-			cfg: fakeClassConfig{
+			cfg: FakeClassConfig{
 				classConfig: map[string]interface{}{
 					"baseURL": "https://proxy.weaviate.dev",
 				},
@@ -231,77 +233,24 @@ func TestClassSettings(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		ic := NewClassSettings(tt.cfg)
+		ic := ent.NewClassSettings(tt.cfg)
 		assert.Equal(t, tt.expectedBaseURL, ic.BaseURL())
-	}
-}
-
-func TestValidateModelVersion(t *testing.T) {
-	type test struct {
-		model    string
-		docType  string
-		version  string
-		possible bool
-	}
-
-	tests := []test{
-		// 001 models
-		{"ada", "text", "001", true},
-		{"ada", "code", "001", true},
-		{"babbage", "text", "001", true},
-		{"babbage", "code", "001", true},
-		{"curie", "text", "001", true},
-		{"curie", "code", "001", true},
-		{"davinci", "text", "001", true},
-		{"davinci", "code", "001", true},
-
-		// 002 models
-		{"ada", "text", "002", true},
-		{"davinci", "text", "002", true},
-		{"ada", "code", "002", false},
-		{"babbage", "text", "002", false},
-		{"babbage", "code", "002", false},
-		{"curie", "text", "002", false},
-		{"curie", "code", "002", false},
-		{"davinci", "code", "002", false},
-
-		// 003
-		{"davinci", "text", "003", true},
-		{"ada", "text", "003", false},
-		{"babbage", "text", "003", false},
-
-		// 004
-		{"davinci", "text", "004", false},
-		{"ada", "text", "004", false},
-		{"babbage", "text", "004", false},
-	}
-
-	for _, test := range tests {
-		name := fmt.Sprintf("model=%s docType=%s version=%s", test.model, test.docType, test.version)
-		t.Run(name, func(t *testing.T) {
-			err := (&classSettings{}).validateModelVersion(test.version, test.model, test.docType)
-			if test.possible {
-				assert.Nil(t, err, "this combination should be possible")
-			} else {
-				assert.NotNil(t, err, "this combination should not be possible")
-			}
-		})
 	}
 }
 
 func TestPickDefaultModelVersion(t *testing.T) {
 	t.Run("ada with text", func(t *testing.T) {
-		version := PickDefaultModelVersion("ada", "text")
+		version := ent.PickDefaultModelVersion("ada", "text")
 		assert.Equal(t, "002", version)
 	})
 
 	t.Run("ada with code", func(t *testing.T) {
-		version := PickDefaultModelVersion("ada", "code")
+		version := ent.PickDefaultModelVersion("ada", "code")
 		assert.Equal(t, "001", version)
 	})
 
 	t.Run("with curie", func(t *testing.T) {
-		version := PickDefaultModelVersion("curie", "text")
+		version := ent.PickDefaultModelVersion("curie", "text")
 		assert.Equal(t, "001", version)
 	})
 }
